refactor: moved upload_file to api
